Trait wry::application::platform::macos::EventLoopExtMacOS

pub trait EventLoopExtMacOS {
    // Required methods
    fn set_activation_policy(&mut self, activation_policy: ActivationPolicy);
    fn enable_default_menu_creation(&mut self, enable: bool);
    fn set_activate_ignoring_other_apps(&mut self, ignore: bool);
}

Required Methods

fn set_activation_policy(&mut self, activation_policy: ActivationPolicy)

Sets the activation policy for the application. It is set to NSApplicationActivationPolicyRegular by default.

This function only takes effect if it’s called before calling run or run_return. To set the activation policy after that, use EventLoopWindowTargetExtMacOS::set_activation_policy_at_runtime.

fn enable_default_menu_creation(&mut self, enable: bool)

Used to prevent a default menubar menu from getting created

The default menu creation is enabled by default.

This function only takes effect if it’s called before calling run or run_return

fn set_activate_ignoring_other_apps(&mut self, ignore: bool)

Used to prevent the application from automatically activating when launched if another application is already active

The default behavior is to ignore other applications and activate when launched.

This function only takes effect if it’s called before calling run or run_return