Trait wry::application::platform::macos::EventLoopWindowTargetExtMacOS

pub trait EventLoopWindowTargetExtMacOS {
    fn hide_application(&self);
    fn show_application(&self);
    fn hide_other_applications(&self);
    fn set_activation_policy_at_runtime(
        &self,
        activation_policy: ActivationPolicy
    );
}

Additional methods on EventLoopWindowTarget that are specific to macOS.

Required Methods

fn hide_application(&self)

Hide the entire application. In most applications this is typically triggered with Command-H.

fn show_application(&self)

Show the entire application.

fn hide_other_applications(&self)

Hide the other applications. In most applications this is typically triggered with Command+Option-H.

fn set_activation_policy_at_runtime(&self, activation_policy: ActivationPolicy)

Sets the activation policy for the application. It is set to NSApplicationActivationPolicyRegular by default.

To set the activation policy before the app starts running, see EventLoopExtMacOS::set_activation_policy.