Struct wry::application::window::WindowAttributes

pub struct WindowAttributes {

    pub inner_size: Option<Size>,
    pub min_inner_size: Option<Size>,
    pub max_inner_size: Option<Size>,
    pub position: Option<Position>,
    pub resizable: bool,
    pub minimizable: bool,
    pub maximizable: bool,
    pub closable: bool,
    pub fullscreen: Option<Fullscreen>,
    pub title: String,
    pub maximized: bool,
    pub visible: bool,
    pub transparent: bool,
    pub decorations: bool,
    pub always_on_top: bool,
    pub always_on_bottom: bool,
    pub window_icon: Option<Icon>,
    pub window_menu: Option<Menu>,
    pub preferred_theme: Option<Theme>,
    pub focused: bool,
    pub content_protection: bool,
    pub visible_on_all_workspaces: bool,

}

Attributes to use when creating a window.

Fields
inner_size: Option<Size>

The dimensions of the window. If this is None, some platform-specific dimensions will be used.

The default is None.

min_inner_size: Option<Size>

The minimum dimensions a window can be, If this is None, the window will have no minimum dimensions (aside from reserved).

The default is None.

max_inner_size: Option<Size>

The maximum dimensions a window can be, If this is None, the maximum will have no maximum or will be set to the primary monitor’s dimensions by the platform.

The default is None.

position: Option<Position>

The desired position of the window. If this is None, some platform-specific position will be chosen.

The default is None.

Platform-specific

macOS: The top left corner position of the window content, the window’s “inner” position. The window title bar will be placed above it. The window will be positioned such that it fits on screen, maintaining set inner_size if any. If you need to precisely position the top left corner of the whole window you have to use Window::set_outer_position after creating the window.

Windows: The top left corner position of the window title bar, the window’s “outer” position. There may be a small gap between this position and the window due to the specifics of the Window Manager.

Linux: The top left corner of the window, the window’s “outer” position.

Linux(Wayland): Unsupported.

Others: Ignored.

See Window::set_outer_position.

resizable: bool

Whether the window is resizable or not.

The default is true.

minimizable: bool

Whether the window is minimizable or not.

The default is true.

See Window::set_minimizable for details.

maximizable: bool

Whether the window is maximizable or not.

The default is true.

See Window::set_maximizable for details.

closable: bool

Whether the window is closable or not.

The default is true.

See Window::set_closable for details.

fullscreen: Option<Fullscreen>

Whether the window should be set as fullscreen upon creation.

The default is None.

title: String

The title of the window in the title bar.

The default is "tao window".

maximized: bool

Whether the window should be maximized upon creation.

The default is false.

visible: bool

Whether the window should be immediately visible upon creation.

The default is true.

transparent: bool

Whether the the window should be transparent. If this is true, writing colors with alpha values different than 1.0 will produce a transparent window.

The default is false.

decorations: bool

Whether the window should have borders and bars.

The default is true.

always_on_top: bool

Whether the window should always be on top of other windows.

The default is false.

always_on_bottom: bool

Whether the window should always be on bottom of other windows.

The default is false.

window_icon: Option<Icon>

The window icon.

The default is None.

window_menu: Option<Menu>

The window menu.

The default is None.

preferred_theme: Option<Theme>focused: bool

Whether the window should be initially focused or not.

Platform-specific:

Android / iOS: Unsupported.

content_protection: bool

Prevents the window contents from being captured by other apps.

Platform-specific

iOS / Android / Linux: Unsupported.

visible_on_all_workspaces: bool

Sets whether the window should be visible on all workspaces.

Platform-specific

iOS / Android / Windows: Unsupported.