Struct wry::application::window::WindowBuilder

pub struct WindowBuilder {
    pub window: WindowAttributes,
    /* private fields */
}

Object that allows you to build windows.

Fields
window: WindowAttributes

The attributes to use to create the window.

Implementations

impl WindowBuilder

pub fn new() -> WindowBuilder

Initializes a new WindowBuilder with default values.

pub fn with_inner_size<S>(self, size: S) -> WindowBuilderwhere S: Into<Size>,

Requests the window to be of specific dimensions.

See Window::set_inner_size for details.

pub fn with_min_inner_size<S>(self, min_size: S) -> WindowBuilderwhere S: Into<Size>,

Sets a minimum dimension size for the window.

See Window::set_min_inner_size for details.

pub fn with_max_inner_size<S>(self, max_size: S) -> WindowBuilderwhere S: Into<Size>,

Sets a maximum dimension size for the window.

See Window::set_max_inner_size for details.

pub fn with_position<P>(self, position: P) -> WindowBuilderwhere P: Into<Position>,

Sets a desired initial position for the window.

See WindowAttributes::position for details.

pub fn with_resizable(self, resizable: bool) -> WindowBuilder

Sets whether the window is resizable or not.

See Window::set_resizable for details.

pub fn with_minimizable(self, minimizable: bool) -> WindowBuilder

Sets whether the window is minimizable or not.

See Window::set_minimizable for details.

pub fn with_maximizable(self, maximizable: bool) -> WindowBuilder

Sets whether the window is maximizable or not.

See Window::set_maximizable for details.

pub fn with_closable(self, closable: bool) -> WindowBuilder

Sets whether the window is closable or not.

See Window::set_closable for details.

pub fn with_title<T>(self, title: T) -> WindowBuilderwhere T: Into<String>,

Requests a specific title for the window.

See Window::set_title for details.

pub fn with_menu(self, menu: MenuBar) -> WindowBuilder

Requests a specific menu for the window.

See Window::set_menu for details.

pub fn with_fullscreen(self, fullscreen: Option<Fullscreen>) -> WindowBuilder

Sets the window fullscreen state.

See Window::set_fullscreen for details.

pub fn with_maximized(self, maximized: bool) -> WindowBuilder

Requests maximized mode.

See Window::set_maximized for details.

pub fn with_visible(self, visible: bool) -> WindowBuilder

Sets whether the window will be initially hidden or visible.

See Window::set_visible for details.

pub fn with_transparent(self, transparent: bool) -> WindowBuilder

Sets whether the background of the window should be transparent.

pub fn with_decorations(self, decorations: bool) -> WindowBuilder

Sets whether the window should have a border, a title bar, etc.

See Window::set_decorations for details.

pub fn with_always_on_bottom(self, always_on_bottom: bool) -> WindowBuilder

Sets whether or not the window will always be below other windows.

See Window::set_always_on_bottom for details.

pub fn with_always_on_top(self, always_on_top: bool) -> WindowBuilder

Sets whether or not the window will always be on top of other windows.

See Window::set_always_on_top for details.

pub fn with_window_icon(self, window_icon: Option<Icon>) -> WindowBuilder

Sets the window icon.

See Window::set_window_icon for details.

pub fn with_theme(self, theme: Option<Theme>) -> WindowBuilder

Forces a theme or uses the system settings if None was provided.

pub fn with_focused(self, focused: bool) -> WindowBuilder

Whether the window will be initially focused or not.

Platform-specific:

Android / iOS: Unsupported.

pub fn with_content_protection(self, protected: bool) -> WindowBuilder

Prevents the window contents from being captured by other apps.

Platform-specific

iOS / Android / Linux: Unsupported.

pub fn with_visible_on_all_workspaces(self, visible: bool) -> WindowBuilder

Sets whether the window should be visible on all workspaces.

Platform-specific

iOS / Android / Windows: Unsupported.

pub fn build<T>( self, window_target: &EventLoopWindowTarget<T> ) -> Result<Window, OsError>where T: 'static,

Builds the window.

Possible causes of error include denied permission, incompatible system, and lack of memory.