wry

Trait WebViewBuilderExtUnix

pub trait WebViewBuilderExtUnix<'a> {
    // Required method
    fn new_gtk<W>(widget: &'a W) -> Self
       where W: IsA<Container>;
}

Required Methods

fn new_gtk<W>(widget: &'a W) -> Self
where W: IsA<Container>,

Create the webview inside a GTK container widget, such as GTK window.

If the container is gtk::Box, it is added using Box::pack_start(webview, true, true, 0).

If the container is gtk::Fixed, its size request will be set using the (width, height) bounds passed in and will be added to the container using Fixed::put using the (x, y) bounds passed in.

For all other containers, it will be added using gtk::prelude::ContainerExt::add

Panics:

Panics if gtk::init was not called in this thread.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.