Struct wry::webview::WebViewAttributes

pub struct WebViewAttributes {

    pub user_agent: Option<String>,
    pub visible: bool,
    pub transparent: bool,
    pub url: Option<Url>,
    pub zoom_hotkeys_enabled: bool,
    pub html: Option<String>,
    pub initialization_scripts: Vec<String>,
    pub custom_protocols: Vec<(String, Box<dyn Fn(&HttpRequest) -> Result<HttpResponse>>)>,
    pub ipc_handler: Option<Box<dyn Fn(&Window, String)>>,
    pub file_drop_handler: Option<Box<dyn Fn(&Window, FileDropEvent) -> bool>>,
    pub navigation_handler: Option<Box<dyn Fn(String) -> bool>>,
    pub new_window_req_handler: Option<Box<dyn Fn(String) -> bool>>,
    pub clipboard: bool,
    pub devtools: bool,

}

Fields
user_agent: Option<String>

Whether the WebView should have a custom user-agent.

visible: bool

Whether the WebView window should be visible.

transparent: bool

Whether the WebView should be transparent. Not supported on Windows 7.

url: Option<Url>

Whether load the provided URL to WebView.

zoom_hotkeys_enabled: bool

Whether page zooming by hotkeys is enabled

Platform-specific

macOS / Linux / Android / iOS: Unsupported

html: Option<String>

Whether load the provided html string to WebView. This will be ignored if the url is provided.

Warning

The loaded from html string will have different Origin on different platforms. And servers which enforce CORS will need to add exact same Origin header in Access-Control-Allow-Origin if you wish to send requests with native fetch and XmlHttpRequest APIs. Here are the different Origin headers across platforms:

macOS: http://localhost

Linux: http://localhost

Windows: null

Android: Not supported

iOS: Not supported

initialization_scripts: Vec<String>

Initialize javascript code when loading new pages. When webview load a new page, this initialization code will be executed. It is guaranteed that code is executed before window.onload.

custom_protocols: Vec<(String, Box<dyn Fn(&HttpRequest) -> Result<HttpResponse>>)>

Register custom file loading protocols with pairs of scheme uri string and a handling closure.

The closure takes a url string slice, and returns a two item tuple of a vector of bytes which is the content and a mimetype string of the content.

Warning

Pages loaded from custom protocol will have different Origin on different platforms. And servers which enforce CORS will need to add exact same Origin header in Access-Control-Allow-Origin if you wish to send requests with native fetch and XmlHttpRequest APIs. Here are the different Origin headers across platforms:

macOS: <scheme_name>://<path> (so it will be wry://examples in custom_protocol example)

Linux: Though it’s same as macOS, there’s a bug that Origin header in the request will be empty. So the only way to pass the server is setting Access-Control-Allow-Origin: *.

Windows: https://<scheme_name>.<path> (so it will be https://wry.examples in custom_protocol example)

Android: Custom protocol on Android is fixed to https://tauri.wry/ due to its design and our approach to use it. On Android, We only handle the scheme name and ignore the closure. So when you load the url like wry://assets/index.html, it will become https://tauri.wry/assets/index.html. Android has assets and resource path finder to locate your files in those directories. For more information, see Loading in-app content page.

iOS: Same as macOS. To get the path of your assets, you can call CFBundle::resources_path. So url like wry://assets/index.html could get the html file in assets directory.

ipc_handler: Option<Box<dyn Fn(&Window, String)>>

Set the IPC handler to receive the message from Javascript on webview to host Rust code. The message sent from webview should call window.ipc.postMessage("insert_message_here");.

Both functions return promises but notify() resolves immediately.

file_drop_handler: Option<Box<dyn Fn(&Window, FileDropEvent) -> bool>>

Set a handler closure to process incoming FileDropEvent of the webview.

Blocking OS Default Behavior

Return true in the callback to block the OS’ default behavior of handling a file drop.

Note, that if you do block this behavior, it won’t be possible to drop files on <input type="file"> forms. Also note, that it’s not possible to manually set the value of a <input type="file"> via JavaScript for security reasons.

navigation_handler: Option<Box<dyn Fn(String) -> bool>>

Set a navigation handler to decide if incoming url is allowed to navigate.

The closure take a String parameter as url and return bool to determine the url. True is allow to nivagate and false is not.

new_window_req_handler: Option<Box<dyn Fn(String) -> bool>>

Set a new window handler to decide if incoming url is allowed to open in a new window.

The closure take a String parameter as url and return bool to determine the url. True is allow to nivagate and false is not.

clipboard: bool

Enables clipboard access for the page rendered on Linux and Windows.

macOS doesn’t provide such method and is always enabled by default. But you still need to add menu item accelerators to use shortcuts.

devtools: bool

Enable web inspector which is usually called dev tool.

Note this only enables dev tool to the webview. To open it, you can call WebView::open_devtools, or right click the page and open it from the context menu.

Platform-specific

macOS: This will call private functions on macOS. It’s still enabled if set in debug build on mac, but requires devtools feature flag to actually enable it in release build.

Android: Open chrome://inspect/#devices in Chrome to get the devtools window. Wry’s WebView devtools API isn’t supported on Android.

iOS: Open Safari > Develop > [Your Device Name] > [Your WebView] to get the devtools window.