Struct wry::webview::WebViewBuilder

pub struct WebViewBuilder<'a> {
    pub webview: WebViewAttributes,
    /* private fields */
}

Builder type of WebView.

WebViewBuilder / WebView are the basic building blocks to constrcut WebView contents and scripts for those who prefer to control fine grained window creation and event handling. WebViewBuilder privides ability to setup initialization before web engine starts.

Fields
webview: WebViewAttributes
Implementations

impl<'a> WebViewBuilder<'a>

pub fn new(window: Window) -> Result<Self>

Create WebViewBuilder from provided Window.

pub fn with_transparent(self, transparent: bool) -> Self

Sets whether the WebView should be transparent. Not supported on Windows 7.

pub fn with_visible(self, visible: bool) -> Self

Sets whether the WebView should be transparent.

pub fn with_initialization_script(self, js: &str) -> Self

Initialize javascript code when loading new pages. When webview load a new page, this initialization code will be executed. It is guaranteed that code is executed before window.onload.

pub fn with_custom_protocol<F>(self, name: String, handler: F) -> Self where
F: Fn(&HttpRequest) -> Result<HttpResponse> + 'static,

Register custom file loading protocols with pairs of scheme uri string and a handling closure.

The closure takes a url string slice, and returns a two item tuple of a vector of bytes which is the content and a mimetype string of the content.

Warning

Pages loaded from custom protocol will have different Origin on different platforms. And servers which enforce CORS will need to add exact same Origin header in Access-Control-Allow-Origin if you wish to send requests with native fetch and XmlHttpRequest APIs. Here are the different Origin headers across platforms:

macOS: <scheme_name>://<path> (so it will be wry://examples in custom_protocol example)

Linux: Though it’s same as macOS, there’s a bug that Origin header in the request will be empty. So the only way to pass the server is setting Access-Control-Allow-Origin: *.

Windows: https://<scheme_name>.<path> (so it will be https://wry.examples in custom_protocol example)

Android: Custom protocol on Android is fixed to https://tauri.wry/ due to its design and our approach to use it. On Android, We only handle the scheme name and ignore the closure. So when you load the url like wry://assets/index.html, it will become https://tauri.wry/assets/index.html. Android has assets and resource path finder to locate your files in those directories. For more information, see Loading in-app content page.

iOS: Same as macOS. To get the path of your assets, you can call CFBundle::resources_path. So url like wry://assets/index.html could get the html file in assets directory.

pub fn with_ipc_handler<F>(self, handler: F) -> Self where
F: Fn(&Window, String) + 'static,

Set the IPC handler to receive the message from Javascript on webview to host Rust code. The message sent from webview should call window.ipc.postMessage("insert_message_here");.

pub fn with_file_drop_handler<F>(self, handler: F) -> Self where
F: Fn(&Window, FileDropEvent) -> bool + 'static,

Set a handler closure to process incoming FileDropEvent of the webview.

Blocking OS Default Behavior

Return true in the callback to block the OS’ default behavior of handling a file drop.

Note, that if you do block this behavior, it won’t be possible to drop files on <input type="file"> forms. Also note, that it’s not possible to manually set the value of a <input type="file"> via JavaScript for security reasons.

pub fn with_url(self, url: &str) -> Result<Self>

Load the provided URL when the builder calling WebViewBuilder::build to create the WebView. The provided URL must be valid.

pub fn with_html(self, html: impl Into<String>) -> Result<Self>

Load the provided HTML string when the builder calling WebViewBuilder::build to create the WebView. This will be ignored if url is already provided.

Warning

The Page loaded from html string will have different Origin on different platforms. And servers which enforce CORS will need to add exact same Origin header in Access-Control-Allow-Origin if you wish to send requests with native fetch and XmlHttpRequest APIs. Here are the different Origin headers across platforms:

macOS: http://localhost

Linux: http://localhost

Windows: null

Android: Not supported

iOS: Not supported

pub fn with_web_context(self, web_context: &'a mut WebContext) -> Self

Set the web context that can share with multiple WebViews.

pub fn with_user_agent(self, user_agent: &str) -> Self

Set a custom user-agent for the WebView.

pub fn with_devtools(self, devtools: bool) -> Self

Enable web inspector which is usually called dev tool.

Note this only enables dev tool to the webview. To open it, you can call WebView::open_devtools, or right click the page and open it from the context menu.

Platform-specific

macOS: This will call private functions on macOS. It’s still enabled if set in debug build on mac, but requires devtools feature flag to actually enable it in release build.

Android: Open chrome://inspect/#devices in Chrome to get the devtools window. Wry’s WebView devtools API isn’t supported on Android.

iOS: Open Safari > Develop > [Your Device Name] > [Your WebView] to get the devtools window.

pub fn with_hotkeys_zoom(self, zoom: bool) -> Self

Whether page zooming by hotkeys or gestures is enabled

Platform-specific

macOS / Linux / Android / iOS: Unsupported

pub fn with_navigation_handler(
self,
callback: impl Fn(String) -> bool + 'static
) -> Self

Set a navigation handler to decide if incoming url is allowed to navigate.

The closure takes a String parameter as url and return bool to determine the url. True is allowed to nivagate and false is not.

pub fn with_new_window_req_handler(
self,
callback: impl Fn(String) -> bool + 'static
) -> Self

Set a new window request handler to decide if incoming url is allowed to be opened.

The closure takes a String parameter as url and return bool to determine if the url can be opened in a new window. Returning true will open the url in a new window, whilst returning false will neither open a new window nor allow any navigation.

pub fn build(self) -> Result<WebView>

Consume the builder and create the WebView.

Platform-specific behavior:

Unix: This method must be called in a gtk thread. Usually this means it should be called in the same thread with the EventLoop you create.