Module wry::application

Re-exported Tao APIs

This module re-export tao APIs for user to create application windows. To learn more about how to use tao, please see its documentation.

Modules

accelerator

Accelerators describe keyboard shortcuts defined by the application.

clipboard

The Clipboard struct and associated types.

dpi

UI scaling is important, so read the docs for this module if you don’t want to be confused.

error

The Error struct and associated types.

event

The Event enum and assorted supporting types.

event_loop

The EventLoop struct and assorted supporting types, including ControlFlow.

global_shortcut

UNSTABLE – The GlobalShortcut struct and associated types.

keyboard

UNSTABLE – Types related to the keyboard.

menu

UNSTABLE – The Menu struct and associated types.

monitor

Types useful for interacting with a user’s monitors.

platform

Contains traits with platform-specific methods in them.

window

The Window struct and associated types.

Structs

TrayId

Whenever you receive an event arising from a particular tray, this event contains a TrayId which identifies its origin.