Trait wry::WebViewBuilderExtUnix

pub trait WebViewBuilderExtUnix<'a> {
    // Required methods
    fn build_gtk<W>(self, widget: &'a W) -> Result<WebView>
       where W: IsA<Container>;
    fn with_extension_path(self, path: impl Into<PathBuf>) -> Self;
}

Required Methods

fn build_gtk<W>(self, widget: &'a W) -> Result<WebView>
where W: IsA<Container>,

Consume the builder and create the webview inside a GTK container widget, such as GTK window.

If the container is gtk::Box, it is added using Box::pack_start(webview, true, true, 0).

If the container is gtk::Fixed, its size request will be set using the (width, height) bounds passed in and will be added to the container using Fixed::put using the (x, y) bounds passed in.

For all other containers, it will be added using gtk::prelude::ContainerExt::add

Panics:

Panics if gtk::init was not called in this thread.

fn with_extension_path(self, path: impl Into<PathBuf>) -> Self

Set the path from which to load extensions from.

Object Safety

This trait is not object safe.