Struct wry::webview::WebViewAttributes

pub struct WebViewAttributes {

    pub user_agent: Option<String>,
    pub visible: bool,
    pub transparent: bool,
    pub background_color: Option<RGBA>,
    pub url: Option<Url>,
    pub headers: Option<HeaderMap>,
    pub zoom_hotkeys_enabled: bool,
    pub html: Option<String>,
    pub initialization_scripts: Vec<String>,
    pub custom_protocols: Vec<(String, Box<dyn Fn(&Request<Vec<u8>>) -> Result<Response<Cow<'static, [u8]>>>>)>,
    pub ipc_handler: Option<Box<dyn Fn(&Window, String)>>,
    pub file_drop_handler: Option<Box<dyn Fn(&Window, FileDropEvent) -> bool>>,
    pub navigation_handler: Option<Box<dyn Fn(String) -> bool>>,
    pub download_started_handler: Option<Box<dyn FnMut(String, &mut PathBuf) -> bool>>,
    pub download_completed_handler: Option<Rc<dyn Fn(String, Option<PathBuf>, bool) + 'static>>,
    pub new_window_req_handler: Option<Box<dyn Fn(String) -> bool>>,
    pub clipboard: bool,
    pub devtools: bool,
    pub accept_first_mouse: bool,
    pub back_forward_navigation_gestures: bool,
    pub document_title_changed_handler: Option<Box<dyn Fn(&Window, String)>>,

}

Fields
user_agent: Option<String>

Whether the WebView should have a custom user-agent.

visible: bool

Whether the WebView window should be visible.

transparent: bool

Whether the WebView should be transparent.

Platform-specific:

Windows 7: Not supported.

background_color: Option<RGBA>

Specify the webview background color. This will be ignored if transparent is set to true.

The color uses the RGBA format.

Platform-specific:

macOS / iOS: Not implemented.

Windows:

On Windows 7, transparency is not supported and the alpha value will be ignored.

On Windows higher than 7: translucent colors are not supported so any alpha value other than 0 will be replaced by 255

url: Option<Url>

Whether load the provided URL to WebView.

headers: Option<HeaderMap>

Headers used when loading the requested url.

zoom_hotkeys_enabled: bool

Whether page zooming by hotkeys is enabled

Platform-specific

macOS / Linux / Android / iOS: Unsupported

html: Option<String>

Whether load the provided html string to WebView. This will be ignored if the url is provided.

Warning

The loaded from html string will have different Origin on different platforms. And servers which enforce CORS will need to add exact same Origin header in Access-Control-Allow-Origin if you wish to send requests with native fetch and XmlHttpRequest APIs. Here are the different Origin headers across platforms:

macOS: http://localhost

Linux: http://localhost

Windows: null

Android: Not supported

iOS: Not supported

initialization_scripts: Vec<String>

Initialize javascript code when loading new pages. When webview load a new page, this initialization code will be executed. It is guaranteed that code is executed before window.onload.

Platform-specific

Android: The Android WebView does not provide an API for initialization scripts, so we prepend them to each HTML head. They are only implemented on custom protocol URLs.

custom_protocols: Vec<(String, Box<dyn Fn(&Request<Vec<u8>>) -> Result<Response<Cow<'static, [u8]>>>>)>

Register custom file loading protocols with pairs of scheme uri string and a handling closure.

The closure takes a Request and returns a Response.

Warning

Pages loaded from custom protocol will have different Origin on different platforms. And servers which enforce CORS will need to add exact same Origin header in Access-Control-Allow-Origin if you wish to send requests with native fetch and XmlHttpRequest APIs. Here are the different Origin headers across platforms:

macOS: <scheme_name>://<path> (so it will be wry://examples in custom_protocol example)

Linux: Though it’s same as macOS, there’s a bug that Origin header in the request will be empty. So the only way to pass the server is setting Access-Control-Allow-Origin: *.

Windows: https://<scheme_name>.<path> (so it will be https://wry.examples in custom_protocol example)

Android: Custom protocol on Android is fixed to https://tauri.wry/ due to its design and our approach to use it. On Android, We only handle the scheme name and ignore the closure. So when you load the url like wry://assets/index.html, it will become https://tauri.wry/assets/index.html. Android has assets and resource path finder to locate your files in those directories. For more information, see Loading in-app content page.

iOS: Same as macOS. To get the path of your assets, you can call CFBundle::resources_path. So url like wry://assets/index.html could get the html file in assets directory.

ipc_handler: Option<Box<dyn Fn(&Window, String)>>

Set the IPC handler to receive the message from Javascript on webview to host Rust code. The message sent from webview should call window.ipc.postMessage("insert_message_here");.

Both functions return promises but notify() resolves immediately.

file_drop_handler: Option<Box<dyn Fn(&Window, FileDropEvent) -> bool>>

Set a handler closure to process incoming FileDropEvent of the webview.

Blocking OS Default Behavior

Return true in the callback to block the OS’ default behavior of handling a file drop.

Note, that if you do block this behavior, it won’t be possible to drop files on <input type="file"> forms. Also note, that it’s not possible to manually set the value of a <input type="file"> via JavaScript for security reasons.

navigation_handler: Option<Box<dyn Fn(String) -> bool>>

Set a navigation handler to decide if incoming url is allowed to navigate.

The closure take a String parameter as url and return bool to determine the url. True is allow to navigate and false is not.

download_started_handler: Option<Box<dyn FnMut(String, &mut PathBuf) -> bool>>

Set a download started handler to manage incoming downloads.

The closure takes two parameters - the first is a String representing the url being downloaded from and and the second is a mutable PathBuf reference that (possibly) represents where the file will be downloaded to. The latter parameter can be used to set the download location by assigning a new path to it - the assigned path must be absolute. The closure returns a bool to allow or deny the download.

download_completed_handler: Option<Rc<dyn Fn(String, Option<PathBuf>, bool) + 'static>>

Sets a download completion handler to manage downloads that have finished.

The closure is fired when the download completes, whether it was successful or not. The closure takes a String representing the URL of the original download request, an Option<PathBuf> potentially representing the filesystem path the file was downloaded to, and a bool indicating if the download succeeded. A value of None being passed instead of a PathBuf does not necessarily indicate that the download did not succeed, and may instead indicate some other failure - always check the third parameter if you need to know if the download succeeded.

Platform-specific:

macOS: The second parameter indicating the path the file was saved to is always empty, due to API limitations.

new_window_req_handler: Option<Box<dyn Fn(String) -> bool>>

Set a new window handler to decide if incoming url is allowed to open in a new window.

The closure take a String parameter as url and return bool to determine the url. True is allow to navigate and false is not.

clipboard: bool

Enables clipboard access for the page rendered on Linux and Windows.

macOS doesn’t provide such method and is always enabled by default. But you still need to add menu item accelerators to use shortcuts.

devtools: bool

Enable web inspector which is usually called dev tool.

Note this only enables dev tool to the webview. To open it, you can call WebView::open_devtools, or right click the page and open it from the context menu.

Platform-specific

macOS: This will call private functions on macOS. It’s still enabled if set in debug build on mac, but requires devtools feature flag to actually enable it in release build.

Android: Open chrome://inspect/#devices in Chrome to get the devtools window. Wry’s WebView devtools API isn’t supported on Android.

iOS: Open Safari > Develop > [Your Device Name] > [Your WebView] to get the devtools window.

accept_first_mouse: bool

Whether clicking an inactive window also clicks through to the webview. Default is false.

Platform-specific

This configuration only impacts macOS.

back_forward_navigation_gestures: bool

Indicates whether horizontal swipe gestures trigger backward and forward page navigation.

Platform-specific

This configuration only impacts macOS. Documentation.

document_title_changed_handler: Option<Box<dyn Fn(&Window, String)>>

Set a handler closure to process the change of the webview’s document title.